Revert "Bug 1940240 - Part 5: Enable container-timing web platform tests. r=canova" for causing wpt failures @containertiming-large-container-size.html. IGNORE BAD COMMIT MESSAGES
This reverts commit d51b3f7b4b1f0cb0867f421c79e145f034cb7d36.

Revert "Bug 1940240 - Part 4: Record container-timing paints from the layout paint pipeline. r=layout-reviewers,emilio"

This reverts commit 26a59da95256d3b4be701b1514626358d84108d2.

Revert "Bug 1940240 - Part 3: Implement the Container Timing API behind a pref. r=smaug,webidl,saschanaz"

This reverts commit 51568e3f54156407b174efe2840988b08cac0f7d.

Revert "Bug 1940240 - Part 2: Compute rect and region areas in 64-bit to avoid overflow. r=gfx-reviewers,bradwerth,lsalzman"

This reverts commit c419214bb9ed840765a48696bfe07b551f50a48f.

Revert "Bug 1940240 - Part 1: Centralize reduced-precision timestamp conversion on Performance. r=canova"

This reverts commit 49261c7747c33e09d9ee9ab5c3f6d42ace6b4d93.

+++ gfx/src/nsRegion.cpp
@@ -483,7 +483,7 @@ uint64_t nsRegion::Area() const {
   for (const Band& band : mBands) {
     uint32_t height = band.bottom - band.top;
     for (const Strip& strip : band.mStrips) {
-      area += int64_t(strip.right - strip.left) * height;
+      area += (strip.right - strip.left) * height;
     }
   }
